pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MuhammadAbdullah95", email = "ma2404374@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "openai-agents[voice]>=0.1.0",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "textual>=3.7.0",
]

[project.scripts]
voice-agent = "voice_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
